Make tests compile in Java 11
Java 11 introduces new overloads to the Inflater.setInput, Inflater.inflate,
and Deflater.deflate methods which take a ByteBuffer instead of byte[].
This introduces ambiguity when they're being called with null (as is the case in
our tests). Explicitly casting the null value to byte[] solves the issue
by explicitly telling the compiler which version should be used.